Added tests using jest and supertest
