+++ package.json
@@ -16,9 +16,15 @@
   "devDependencies": {
     "@types/express": "^4.17.12",
     "ts-node-dev": "^1.1.6",
-    "typescript": "^4.3.2"
+    "typescript": "^4.3.2",
+    "supertest": "^6.1.3",
+    "jest": "^27.0.1",
+    "ts-jest": "^27.0.1",
+    "@types/jest": "^26.0.23",
+    "@types/supertest": "^2.0.11"
   },
   "scripts": {
-    "dev": "ts-node-dev --respawn --transpile-only src/index.ts"
+    "dev": "ts-node-dev --respawn --transpile-only src/index.ts",
+    "test": "jest"
   }
 }
