+++ tsconfig.test.json
@@ -1,0 +1,7 @@
+{
+    "extends": "./tsconfig",
+    "compilerOptions": {
+        "target": "ES2019",
+        "allowJs": true
+    }
+}
